+++ example/lib/inputs.dart
@@ -17,6 +17,7 @@ class InputsView extends StatefulWidget {
 }
 
 class _InputsViewState extends State<InputsView> {
+  bool isDisabled = false;
   List<int> selected = [];
   Duration? durationValue = const Duration(
     days: 0,
@@ -25,6 +26,11 @@ class _InputsViewState extends State<InputsView> {
     hours: 0,
   );
   List<String> searchChoices = [];
+  DateTime _selectedDate = DateTime.now();
+  List<DateTime> _selectedDates = [
+    DateTime(2023, 8, 22),
+    DateTime(2023, 8, 26),
+  ];
   ThemedMonth? _selectedMonth = const ThemedMonth(year: 2023, month: Month.july);
   List<ThemedMonth> _selectedMonths = const [
     ThemedMonth(year: 2023, month: Month.july),
@@ -58,7 +64,16 @@ class _InputsViewState extends State<InputsView> {
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
+              ThemedButton(
+                labelText: "Lock / Unlock",
+                onTap: () {
+                  setState(() {
+                    isDisabled = !isDisabled;
+                  });
+                },
+              ),
               ThemedTextInput(
+                disabled: isDisabled,
                 labelText: 'Text Input [Plain]',
                 value: textInputValue,
                 onChanged: (value) {
@@ -91,6 +106,7 @@ class _InputsViewState extends State<InputsView> {
               ),
 
               ThemedTextInput(
+                disabled: isDisabled,
                 labelText: 'Text Input [Combobox]',
                 value: textInputValue2,
                 choices: searchChoices,
@@ -100,7 +116,33 @@ class _InputsViewState extends State<InputsView> {
                   });
                 },
               ),
+              ThemedDatePicker(
+                disabled: isDisabled,
+                labelText: "Date picker",
+                value: _selectedDate,
+                disabledDays: [
+                  DateTime(2023, 8, 20),
+                  DateTime(2023, 8, 26),
+                ],
+                onChanged: (date) {
+                  setState(() {
+                    _selectedDate = date;
+                  });
+                },
+              ),
+              ThemedDateRangePicker(
+                disabled: isDisabled,
+                labelText: "Date range picker",
+                value: _selectedDates,
+                pattern: '%B %d, %Y',
+                onChanged: (dates) {
+                  setState(() {
+                    _selectedDates = dates;
+                  });
+                },
+              ),
               ThemedMonthPicker(
+                disabled: isDisabled,
                 labelText: "Month picker",
                 value: _selectedMonth,
                 minimum: const ThemedMonth(year: 2023, month: Month.january),
@@ -117,6 +159,7 @@ class _InputsViewState extends State<InputsView> {
                 },
               ),
               ThemedMonthRangePicker(
+                disabled: isDisabled,
                 labelText: "Month range picker [Random]",
                 value: _selectedMonths,
                 minimum: const ThemedMonth(year: 2023, month: Month.january),
@@ -133,6 +176,7 @@ class _InputsViewState extends State<InputsView> {
                 },
               ),
               ThemedMonthRangePicker(
+                disabled: isDisabled,
                 labelText: "Month range picker [Consecutive]",
                 value: _selectedMonths2,
                 consecutive: true,
@@ -150,6 +194,7 @@ class _InputsViewState extends State<InputsView> {
                 },
               ),
               ThemedColorPicker(
+                disabled: isDisabled,
                 labelText: 'Color picker',
                 onChanged: (color) {
                   debugPrint("Color: $color");
@@ -157,6 +202,7 @@ class _InputsViewState extends State<InputsView> {
               ),
               Text(durationValue.toString()),
               ThemedDurationInput(
+                disabled: isDisabled,
                 value: durationValue,
                 // visibleValues: const [
                 //   ThemedDurationInputVisibleValues.hours,
@@ -219,81 +265,48 @@ class _InputsViewState extends State<InputsView> {
                   ],
                 ),
               ),
-              Row(
-                children: [
-                  Expanded(
-                    child: ThemedTextInput(
-                      labelText: 'Text Input [Combobox]',
-                      value: textInputValue2,
-                      choices: searchChoices,
-                      onChanged: (value) {
-                        setState(() {
-                          textInputValue2 = value;
-                        });
-                      },
-                      onTap: () {
-                        debugPrint("Combobox tapped");
-                      },
-                    ),
-                  ),
-                  Expanded(
-                    child: ThemedTextInput(
-                      labelText: 'Text Input [Combobox]',
-                      value: textInputValue2,
-                      choices: searchChoices,
-                      onChanged: (value) {
-                        setState(() {
-                          textInputValue2 = value;
-                        });
-                      },
-                      onTap: () {
-                        debugPrint("Combobox tapped");
-                      },
-                    ),
-                  ),
-                ],
-              ),
-              const ThemedDynamicAvatarInput(
+              ThemedDynamicAvatarInput(
+                disabled: isDisabled,
                 labelText: "Dynamic Avatar Input",
               ),
               ThemedIconPicker(
+                disabled: isDisabled,
                 labelText: "Icon picker",
                 onChanged: (icon) {
                   debugPrint("Icon: $icon");
                 },
               ),
               ThemedEmojiPicker(
+                disabled: isDisabled,
                 value: "🍊",
                 labelText: "Emoji picker",
                 onChanged: (emoji) {
                   debugPrint("Emoji: $emoji");
                 },
               ),
-              const ThemedCodeEditor(
+              ThemedCodeEditor(
+                disabled: isDisabled,
                 labelText: "Code Editor",
               ),
               ThemedTextInput(
+                disabled: isDisabled,
                 labelText: "Text Input",
                 prefixText: "prefix.",
                 suffixIcon: MdiIcons.homeVariant,
               ),
-              const ThemedFileInput(
+              ThemedFileInput(
+                disabled: isDisabled,
                 labelText: "File Input",
               ),
               ThemedTextInput(
+                disabled: isDisabled,
                 labelText: "Text Input",
                 prefixText: "prefix.",
                 suffixIcon: MdiIcons.homeVariant,
                 dense: true,
               ),
-              const ThemedDateTimePicker(
-                labelText: "Date Time Picker",
-                showTime: true,
-              ),
-              const ThemedDateTimeRangePicker(
-                labelText: "Date Time range Picker",
-              ),
               ThemedMultiSelectInput<int>(
+                disabled: isDisabled,
                 labelText: "MultiSelect Input",
                 items: choices,
                 value: null,
@@ -305,6 +318,7 @@ class _InputsViewState extends State<InputsView> {
               ),
               ...ThemedCheckboxInputStyle.values.map((style) {
                 return ThemedCheckboxInput(
+                  disabled: isDisabled,
                   labelText: "Checkbox Input with style $style",
                   style: style,
                   value: value,
@@ -316,14 +330,17 @@ class _InputsViewState extends State<InputsView> {
                 );
               }).toList(),
               ThemedSelectInput<int>(
+                disabled: isDisabled,
                 labelText: "Select Input",
                 items: choices,
               ),
               ThemedRadioInput<int>(
+                disabled: isDisabled,
                 labelText: "Radio Input",
                 items: choices,
               ),
               ThemedDualListInput(
+                disabled: isDisabled,
                 labelText: "Dual List Input",
                 items: choices,
               ),