+++ lib/src/inputs/src/pickers/month/range.dart
@@ -19,6 +19,7 @@ class ThemedMonthRangePicker extends StatefulWidget {
   final ThemedMonth? maximum;
   final List<ThemedMonth> disabledMonths;
 
+  /// Creates a [ThemedMonthRangePicker] input.
   const ThemedMonthRangePicker({
     super.key,
 
@@ -64,30 +65,6 @@ class ThemedMonthRangePicker extends StatefulWidget {
     /// but you can submit your own translations using this property. Consider when [LayrzAppLocalizations] is present,
     /// is the default value of this property.
     /// Required translations:
-    /// - `layrz.monthPicker.months.0` (January)
-    /// - `layrz.monthPicker.months.1` (February)
-    /// - `layrz.monthPicker.months.2` (March)
-    /// - `layrz.monthPicker.months.3` (April)
-    /// - `layrz.monthPicker.months.4` (May)
-    /// - `layrz.monthPicker.months.5` (June)
-    /// - `layrz.monthPicker.months.6` (July)
-    /// - `layrz.monthPicker.months.7` (August)
-    /// - `layrz.monthPicker.months.8` (September)
-    /// - `layrz.monthPicker.months.9` (October)
-    /// - `layrz.monthPicker.months.10` (November)
-    /// - `layrz.monthPicker.months.11` (December)
-    /// - `layrz.monthPicker.monthsShort.0` (Jan)
-    /// - `layrz.monthPicker.monthsShort.1` (Feb)
-    /// - `layrz.monthPicker.monthsShort.2` (Mar)
-    /// - `layrz.monthPicker.monthsShort.3` (Apr)
-    /// - `layrz.monthPicker.monthsShort.4` (May)
-    /// - `layrz.monthPicker.monthsShort.5` (Jun)
-    /// - `layrz.monthPicker.monthsShort.6` (Jul)
-    /// - `layrz.monthPicker.monthsShort.7` (Aug)
-    /// - `layrz.monthPicker.monthsShort.8` (Sep)
-    /// - `layrz.monthPicker.monthsShort.9` (Oct)
-    /// - `layrz.monthPicker.monthsShort.10` (Nov)
-    /// - `layrz.monthPicker.monthsShort.11` (Dec)
     /// - `actions.cancel` (Cancel)
     /// - `actions.save` (Save)
     /// - `actions.reset` (Reset)
@@ -95,30 +72,6 @@ class ThemedMonthRangePicker extends StatefulWidget {
     /// - `layrz.monthPicker.back` (Previous year)
     /// - `layrz.monthPicker.next` (Next year)
     this.translations = const {
-      'layrz.monthPicker.months.0': 'January',
-      'layrz.monthPicker.months.1': 'February',
-      'layrz.monthPicker.months.2': 'March',
-      'layrz.monthPicker.months.3': 'April',
-      'layrz.monthPicker.months.4': 'May',
-      'layrz.monthPicker.months.5': 'June',
-      'layrz.monthPicker.months.6': 'July',
-      'layrz.monthPicker.months.7': 'August',
-      'layrz.monthPicker.months.8': 'September',
-      'layrz.monthPicker.months.9': 'October',
-      'layrz.monthPicker.months.10': 'November',
-      'layrz.monthPicker.months.11': 'December',
-      'layrz.monthPicker.monthsShort.0': 'Jan',
-      'layrz.monthPicker.monthsShort.1': 'Feb',
-      'layrz.monthPicker.monthsShort.2': 'Mar',
-      'layrz.monthPicker.monthsShort.3': 'Apr',
-      'layrz.monthPicker.monthsShort.4': 'May',
-      'layrz.monthPicker.monthsShort.5': 'Jun',
-      'layrz.monthPicker.monthsShort.6': 'Jul',
-      'layrz.monthPicker.monthsShort.7': 'Aug',
-      'layrz.monthPicker.monthsShort.8': 'Sep',
-      'layrz.monthPicker.monthsShort.9': 'Oct',
-      'layrz.monthPicker.monthsShort.10': 'Nov',
-      'layrz.monthPicker.monthsShort.11': 'Dec',
       'actions.cancel': 'Cancel',
       'actions.save': 'Save',
       'actions.reset': 'Reset',
@@ -145,6 +98,7 @@ class ThemedMonthRangePicker extends StatefulWidget {
 }
 
 class _ThemedMonthRangePickerState extends State<ThemedMonthRangePicker> {
+  LayrzAppLocalizations? get i18n => LayrzAppLocalizations.of(context);
   late int _focusYear;
   bool get isDark => Theme.of(context).brightness == Brightness.dark;
 
@@ -156,7 +110,7 @@ class _ThemedMonthRangePickerState extends State<ThemedMonthRangePicker> {
     final sortedValues = List<ThemedMonth>.from(widget.value)..sort(_sortMonths);
 
     return sortedValues.map((e) {
-      return "${t('layrz.monthPicker.monthsShort.${e.month.index}')} ${e.year}";
+      return "${DateTime(2023, e.month.index + 1, 1).format(pattern: '%B', i18n: i18n)} ${e.year}";
     }).join(', ');
   }
 
@@ -351,7 +305,7 @@ class _ThemedMonthRangePickerState extends State<ThemedMonthRangePicker> {
                                           },
                                     child: Center(
                                       child: Text(
-                                        t('layrz.monthPicker.months.$index'),
+                                        DateTime(2023, index + 1, 1).format(pattern: '%B', i18n: i18n),
                                         style: Theme.of(context).textTheme.bodyMedium?.copyWith(
                                               color: validateColor(color: decoration.color ?? Colors.white),
                                             ),